requirements.txt
fastapi
uvicorn
dependency-injector
sqlalchemy
psycopg2-binary
pydantic
python-dotenv